third_party/blink/web_tests/fast/forms/textarea/textarea-maxlength-expected.txt

Tests for HTMLTextAreaElement.maxLength behaviors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textArea.maxLength is -1
PASS textArea.maxLength is -1
PASS textArea.maxLength is -1
PASS textArea.maxLength is 1
PASS textArea.maxLength is -1
PASS textArea.maxLength is -1
PASS textArea.maxLength is 1
PASS textArea.maxLength is 256
PASS textArea.getAttribute("maxlength") is "13"
PASS textArea.maxLength = -1 threw exception IndexSizeError: Failed to set the 'maxLength' property on 'HTMLTextAreaElement': The value provided (-1) is not positive or 0..
PASS textArea.getAttribute("maxlength") is "13"
PASS textArea.maxLength = 10 threw exception IndexSizeError: Failed to set the 'maxLength' property on 'HTMLTextAreaElement': The maxLength provided (10) is less than the minimum bound (11)..
PASS textArea.getAttribute("maxlength") is "13"
PASS textArea.maxLength = 11; textArea.getAttribute("maxlength") is "11"
PASS textArea.maxLength is 0
PASS textArea.getAttribute("maxlength") is "0"
PASS textArea.value is "abcd"
PASS textArea.value is "abcde"
PASS textArea.value is "abc"
PASS textArea.value is "abc"
PASS textArea.value is "def"
PASS textArea.value is "abcdef"
PASS textArea.value is "abcde"
PASS textArea.value is "A\nB"
PASS textArea.value is "a\n\n"
PASS textArea.value is "\n\n\n\n\n\n"
Inserts 2 normal characters + a combining letter with 3 code points into a maxlength=3 element.
PASS textArea.value is "ABx"
PASS textArea.value.length is 3
PASS textArea.value is "ABC"
Inserts 2 normal characters + one surrogate pair into a maxlength=3 element
PASS textArea.value is "AB"
PASS textArea.value.length is 2
PASS textArea.value is "ABC"
PASS textArea.value is ""
PASS textArea.value is "ABC"
PASS textArea.value is "ABC"
PASS successfullyParsed is true

TEST COMPLETE
